# backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: GameHub API
  description: GameHub游戏社区平台API文档
  version: 1.0.0
  contact:
    name: GameHub开发团队
    email: dev@gamehub.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: 本地开发服务器
  - url: https://api.gamehub.example.com/api/v1
    description: 生产服务器
...
paths:
...
  /users:
    get:
      tags:
        - 用户
      summary: 获取用户列表（管理员）
      description: 获取所有用户列表，仅管理员可访问
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: 页码
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: 每页数量
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: search
          in: query
          description: 搜索关键词（用户名或邮箱）
          schema:
            type: string
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '401':
          description: 未授权
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{id}:
    get:
      tags:
        - 用户
      summary: 获取用户详情
      description: 根据ID获取用户详情
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 用户ID
          schema:
            type: string
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetailResponse'
        '401':
          description: 未授权
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 用户不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - 用户
      summary: 更新用户信息（管理员）
      description: 更新用户信息，仅管理员可访问
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 用户ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserAdminRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetailResponse'
        '401':
          description: 未授权
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - 用户
      summary: 删除用户（管理员）
      description: 删除用户，仅管理员可访问
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 用户ID
          schema:
            type: string
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: 未授权
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /like/user:
    get:
      tags:
        - 点赞
      summary: 获取用户点赞列表
      description: 获取指定用户的点赞历史
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 获取成功
...
  /admin/users:
    get:
      tags:
        - 管理
      summary: 用户管理列表
      description: 管理端获取用户列表
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: role
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 获取成功
...
  /admin/users/{id}:
    get:
      tags:
        - 管理
      summary: 获取用户详情（管理端）
      description: 管理端获取单个用户详细信息
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 获取成功
    put:
      tags:
        - 管理
      summary: 更新用户（管理端）
      description: 管理端更新用户信息
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserAdminRequest'
      responses:
        '200':
          description: 更新成功
    delete:
      tags:
        - 管理
      summary: 删除用户（管理端）
      description: 管理端删除用户
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 删除成功
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
        username:
          type: string
          example: "john_doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        displayName:
          type: string
          example: "John Doe"
        avatarUrl:
          type: string
          format: uri
          example: "https://example.com/avatar.jpg"
        role:
          type: string
          enum: [user, admin, moderator]
          example: "user"
        emailVerified:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2026-03-31T10:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-04-15T14:30:00Z"
    UpdateUserRequest:
      type: object
      properties:
        displayName:
          type: string
          maxLength: 100
          example: "John Smith"
        avatarUrl:
          type: string
          format: uri
          example: "https://example.com/new-avatar.jpg"
        bio:
          type: string
          maxLength: 500
          example: "游戏爱好者"
    UpdateUserAdminRequest:
      allOf:
        - $ref: '#/components/schemas/UpdateUserRequest'
        - type: object
          properties:
            role:
              type: string
              enum: [user, admin, moderator]
              example: "admin"
            status:
              type: string
              enum: [active, inactive, suspended]
              example: "active"
...
    UserDetailResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/User'
        message:
          type: string
          example: "用户详情获取成功"
    UserListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            users:
              type: array
              items:
                $ref: '#/components/schemas/User'
            pagination:
              $ref: '#/components/schemas/Pagination'
        message:
          type: string
          example: "用户列表获取成功"

    # 游戏相关
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "操作成功"
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "错误描述"
        code:
          type: string
          example: "VALIDATION_ERROR"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: "email"
              message:
                type: string
                example: "邮箱格式不正确"
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
        pages:
          type: integer
          example: 5
        hasNext:
          type: boolean
          example: true
        hasPrev:
          type: boolean
          example: false

    # 新闻相关
...